# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIGEA REST API",
        "description": "API REST del Sistema Integral de Gestión Educativa Académica (SIGEA). Autenticación mediante tokens Sanctum (header `Authorization: Bearer <token>`).",
        "contact": {
            "name": "Equipo SIGEA",
            "email": "contacto@sigea.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/v1/alumnos": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Listar alumnos con filtros",
                "operationId": "60e41cd980b4e8c9d93e1f675839d5a9",
                "parameters": [
                    {
                        "name": "buscar",
                        "in": "query",
                        "description": "Texto a buscar en matrícula, nombre o apellidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estatus",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activo",
                                "baja_temporal",
                                "baja_definitiva",
                                "egresado"
                            ]
                        }
                    },
                    {
                        "name": "carrera",
                        "in": "query",
                        "description": "ID de carrera",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/alumnos/{id}": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Detalle de un alumno",
                "operationId": "f670ebce2262af876604853d043d237a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumno encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión y obtener token",
                "operationId": "b9b760991d94b179960cd7ddc93eedf2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@sigea.local"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token emitido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|a8f2h9j3k..."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales inválidas o cuenta desactivada"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión (revoca el token actual)",
                "operationId": "b66dd8a7864f561822c01bc8339363fb",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuario autenticado",
                "operationId": "c24e88aea4324667c5f4e38549a5298f",
                "responses": {
                    "200": {
                        "description": "Datos del usuario"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/kardex": {
            "get": {
                "tags": [
                    "Kardex"
                ],
                "summary": "Historial académico del alumno autenticado",
                "operationId": "4ea465dbfe28f16a5232222be4620f4b",
                "responses": {
                    "200": {
                        "description": "Historial agrupado por ciclo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "alumno": {
                                            "type": "object"
                                        },
                                        "promedio": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "historial": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El usuario no es alumno"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/kardex/pdf": {
            "get": {
                "tags": [
                    "Kardex"
                ],
                "summary": "Descargar kardex en PDF",
                "operationId": "3507ab58fc779b893787e8cf219d0160",
                "responses": {
                    "200": {
                        "description": "Archivo PDF",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "404": {
                        "description": "El usuario no es alumno"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}